Fixed Compatiblity Issues with Python 3.9 (#18)
According to #17, the code does not work with Python 3.9. I have updated tox to include 3.9 and also 3.12. I dropped 3.8 because it never worked and support for python will end in October 2024.

This PR contains code changes to make it compatible with 3.9. Most of the changes fix the new typing syntax. In api.py, the decorator for re-login needs to be moved to a function instead of a static method.

- Updated testing environment (tox and github).
- Fixed compatibility issues with Python 3.9.
- Added tox path to gitignore.
- Updated README with tox information.