Add Installer Authentication to CLI
This change introduces installer authentication to the CLI, allowing users to authenticate with a master key and service code. It also adds a new --credentials option to specify a file containing authentication credentials and deprecates the existing password file.

The changes include improvements to session handling, argument parsing, and code quality, as well as new tests to verify the new authentication methods and options.

pull request #23